AI_Service/requirements.txt
grpcio
protobuf
redis
openai
python-dotenv